Remove date time from double

+++ src/WebDriver.BiDi/Internal/Json/JsonDateTimeConverter.cs
@@ -9,8 +9,7 @@ internal class JsonDateTimeConverter : JsonConverter<DateTime>
     static readonly DateTime s_epoch = new(1970, 1, 1, 0, 0, 0);
     public override DateTime Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
-        // chrome returns sometime doubles, should be UInt64
-        var unixTime = reader.GetDouble();
+        var unixTime = reader.GetUInt64();
 
         return s_epoch.AddMilliseconds(unixTime);
     }
